fix: rename systemd socket fd handle
Dohnut looks for systemd and launchd file descriptors by the name "dohnut". By default, however, systemd `*.socket` files append a `.socket` suffix to the fd name. Set the name to `dohnut` in the dohnut.socket file to normalise it with launch's example configuration.

See:
https://www.freedesktop.org/software/systemd/man/systemd.socket.html#FileDescriptorName=
"If this setting is not used, the file descriptor name defaults to the name of the socket unit, including its .socket suffix."